services/integration-service/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
pydantic-settings
python-dotenv
redis[hiredis]
httpx